Use WWBOTA Server-Side Events API for spots
This uses SSE for receiving spots from the WWBOTA API, meaning low
latency of spots and reducing bandwidth.

With SSE, each spot is sent as a event in same format as standard API,
rather than a list of spots. Debounce is used to avoid multiple map
updates on intial connection when last 2 hours of spots are sent.